Fix cart detail btn

+++ carts/static/carts/js/cart_check.js
@@ -11,42 +11,47 @@ const delBtn = document.getElementById('del_btn')
 
 delBtn.addEventListener('click', function(e) {
   e.preventDefault()
-  const confirmed = confirm("선택한 상품을 장바구니에서 삭제하시겠습니까?");
-  
-  if (confirmed) {
-    const checkItems = document.querySelectorAll('input[name="item_check"]')
-    const checkedItems = [...checkItems].filter((item) => item.checked)
-    // console.log(checkItems)
-    // console.log(checkedItems)
+  const checkItems = document.querySelectorAll('input[name="item_check"]')
+  const checkedItems = [...checkItems].filter((item) => item.checked)
 
-    if (isAuthenticated) {
-      const productIds = checkedItems.map((item) => item.value)
+  if (checkedItems.length > 0) {
+    const confirmed = confirm("선택한 상품을 장바구니에서 삭제하시겠습니까?");
+    
+    if (confirmed) {
+      // console.log(checkItems)
+      // console.log(checkedItems)
 
-      axios({
-        method: 'POST',
-        url: '/carts/remove_item/',
-        headers: {'X-CSRFToken': csrfToken},
-        data: JSON.stringify({
-          productIds,
+      if (isAuthenticated) {
+        const productIds = checkedItems.map((item) => item.value)
+
+        axios({
+          method: 'POST',
+          url: '/carts/remove_item/',
+          headers: {'X-CSRFToken': csrfToken},
+          data: JSON.stringify({
+            productIds,
+          })
         })
-      })
-        .then(function(response) {
-          window.location.href = '/carts/'
-        })
-    } else {
-      const productIds = checkedItems.map((item) => parseInt(item.value, 10));
+          .then(function(response) {
+            window.location.href = '/carts/'
+          })
+      } else {
+        const productIds = checkedItems.map((item) => parseInt(item.value, 10));
 
-      let cart = JSON.parse(localStorage.getItem('cart'));
+        let cart = JSON.parse(localStorage.getItem('cart'));
 
-      cart = cart.filter((item) => !productIds.includes(item.id));
+        cart = cart.filter((item) => !productIds.includes(item.id));
 
-      const countBadge = document.getElementById('count_badge')
+        const countBadge = document.getElementById('count_badge')
 
-      countBadge.textContent = cart.length
-      
-      localStorage.setItem('cart', JSON.stringify(cart));
+        countBadge.textContent = cart.length
+        
+        localStorage.setItem('cart', JSON.stringify(cart));
 
-      window.location.href = '/carts/'
+        window.location.href = '/carts/'
+      }
+    }
+  } else {
+    alert('선택한 상품이 없습니다.')
   }
-}
 })
